Fix bug in spec and fix up driver
The definition of `new-symbolic-entry` had a typo in it such that new
symbolic entries always had their valid field set to `#f`. The
`new-symbolic-state` function was defined in terms of this function, so
this meant that a new symbolic state consisted of only invalid (empty)
entries. The proof infrastructure uses the `new-symbolic-state` function
as a "forall state, ..." mechanism, so this meant that the proofs (both
functional equivalence and physical equivalence) were only covering the
case where all entries started out as invalid before the operation, in
the case of functional equivalence, or before adversary takeover, in the
case of physical equivalence.

For the correctness proof, this meant that the proof could take
advantage of all entries being invalid, so the proof didn't need to
explore the valid case at all. The proof script didn't directly abuse
this; it happened indirectly due to concretization. Fixing up the
correctness proof actually required a spec change: for the case of the
retrieve operation where the entry was valid, the driver was not waiting
long enough for the circuit to reach a quiescent state.

Because of the way the security proof was written, it was not affected
by this bug: early on, the proof replaced the spec state with a term
that was computed as the abstraction function applied to the circuit
state, and because the circuit FRAM was fully unconstrained and the
abstraction function was written correctly, the proof effectively
covered the case of entries being valid as well: it was proving a more
general property than necessary before the spec bug was fixed. For this
reason, fixing the spec bug didn't really affect the physical
equivalence proof's execution or running time.

This fix impacts verification performance. The following numbers were
obtained on the same machine used for the benchmarks in the Knox paper
(Intel i7-5930K). The "before" row is from the Knox paper's Figure 18.
The "after" row is run with respect to Knox commit
4dd42b5c083f8dc6f2facaa47ba69ff922879dcb. It's not an apples-to-apples
comparison because there have been some small changes and improvements
in the Knox framework in the meantime.

| HSM                 | FE-N | FE-N+C | FE  | FE+C | PE |
|---------------------|------|--------|-----|------|----|
| PIN backup (before) | 1    | 10     | 209 | 962  | 8  |
| PIN backup (after)  | 3    | 5      | 216 | 427  | 8  |